refactoring

+++ views/evaluations.php
@@ -4,13 +4,16 @@ if (!defined('APP_ACCESS')) {
   exit;
 }
 
+$currentSort = isset($_GET['sort']) && $_GET['sort'] === 'desc' ? 'desc' : 'asc';
+
 $evaluations =  $type == "exam" ? $exams : $assigments;
 $resultsPerPage = 10;
 $currentPage = isset($_GET['page']) ? (int)$_GET['page'] : 1;
 $totalResults = count($evaluations->findAll());
+$search = isset($_GET['search']) ? trim($_GET['search']) : '';
 
-list($offset, $paginationHtml) = paginate($totalResults, $resultsPerPage, $currentPage, 'evaluations');
-$evaluationsList = $evaluations->findAllPaginated($offset, $resultsPerPage);
+list($offset, $paginationHtml) = paginate($totalResults, $resultsPerPage, $currentPage, $type);
+$evaluationsList = $evaluations->findAllPaginated($offset, $resultsPerPage, $search);
 ?>
 
 
@@ -20,13 +23,16 @@ $evaluationsList = $evaluations->findAllPaginated($offset, $resultsPerPage);
     <div class="row">
       <div class="row search-input">
         <i class="ri-search-line"></i>
-        <input type="text" placeholder="Search..." name="search" id="search" />
+        <form id="searchFom"><input type="text" placeholder="Search..." name="search" id="search" /></form>
       </div>
       <button class="btn"><i class="ri-equalizer-line"></i></button>
-      <button class="btn"><i class="ri-sort-desc"></i></button>
-      <button type="button" class="add-btn btn">
-        <i class="ri-add-line"></i>
-      </button>
+      <button class="btn" id="sortToggle"><i class="ri-sort-desc"></i></button>
+      <?php if ($isAdmin) { ?>
+        <button class="btn" id="sortToggle">
+          <i class="ri-sort-<?= $currentSort ?>"></i>
+        </button>
+      <?php } ?>
+
     </div>
   </div>
 
@@ -36,24 +42,31 @@ $evaluationsList = $evaluations->findAllPaginated($offset, $resultsPerPage);
         <th>Title</th>
         <th>Start Time</th>
         <th>End Time</th>
-        <th>Actions</th>
+        <?php if ($isAdmin) { ?><th>Actions</th><?php } ?>
       </tr>
     </thead>
     <tbody>
       <?php foreach ($evaluationsList as $evaluation): ?>
-        <tr role="link" tabindex="0" class="table-row" onclick="window.location='/evaluations/<?= $evaluation['id'] ?>'">
+        <tr class="table-row">
           <td><?= htmlspecialchars($evaluation['title']) ?></td>
-          <td><?= htmlspecialchars($evaluation['startTime']); ?></td>
-          <td><?= htmlspecialchars($evaluation['endTime']); ?></td>
-          <td class="row" id="actions">
-            <button
-              class="btn edit-btn"
-              data-id="<?= $evaluation['id'] ?>"
-              data-title="<?= htmlspecialchars($evaluation['title']) ?>"
-              data-start-time="<?= htmlspecialchars($evaluation['startTime']) ?>"
-              data-end-time="<?= htmlspecialchars($evaluation['endTime']) ?>"><i class="ri-edit-box-line"></i></button>
-            <a class="btn edit-btn" href="/delete_evaluations?id=<?= $evaluation['id'] ?>"><i class="ri-delete-bin-6-line"></i></a>
+          <td>
+            <?= htmlspecialchars((new DateTime($evaluation['startTime']))->format('d/m/Y - H:i')); ?>
           </td>
+          <td>
+            <?= htmlspecialchars((new DateTime($evaluation['endTime']))->format('d/m/Y - H:i')); ?>
+          </td>
+
+          <?php if ($isAdmin) { ?>
+            <td class="row" id="actions">
+              <button
+                class="btn edit-btn"
+                data-id="<?= $evaluation['id'] ?>"
+                data-title="<?= htmlspecialchars($evaluation['title']) ?>"
+                data-start-time="<?= htmlspecialchars($evaluation['startTime']) ?>"
+                data-end-time="<?= htmlspecialchars($evaluation['endTime']) ?>"><i class="ri-edit-box-line"></i></button>
+              <a class="btn edit-btn" href="/delete_evaluations?id=<?= $evaluation['id'] ?>"><i class="ri-delete-bin-6-line"></i></a>
+            </td>
+          <?php } ?>
         </tr>
       <?php endforeach; ?>
     </tbody>
@@ -107,5 +120,28 @@ $evaluationsList = $evaluations->findAllPaginated($offset, $resultsPerPage);
       document.getElementById('start-time').value = this.dataset.startTime;
       document.getElementById('end-time').value = this.dataset.endTime;
     });
+  });
+
+  document.getElementById("sortToggle").addEventListener("click", () => {
+    const url = new URL(window.location);
+    const currentSort = url.searchParams.get("sort") || "asc";
+    const nextSort = currentSort === "asc" ? "desc" : "asc";
+
+    url.searchParams.set("sort", nextSort);
+    window.location = url;
+  });
+
+  console.log(document.getElementById("search"))
+
+  document.getElementById("search").addEventListener("keydown", function(event) {
+    if (event.key === "Enter") {
+      alert("pressed")
+      event.preventDefault();
+
+      const url = new URL(window.location);
+      url.searchParams.set("search", this.value);
+
+      window.location = url;
+    }
   });
 </script>
